+++ tools/Azure.Mcp.Tools.AppConfig/src/Options/AppConfigOptionDefinitions.cs
@@ -12,50 +12,40 @@ public static class AppConfigOptionDefinitions
     public const string ContentTypeName = "content-type";
     public const string TagsName = "tags";
     public const string LockName = "lock";
+    public const string KeyFilterName = "key-filter";
+    public const string LabelFilterName = "label-filter";
 
-    public static readonly Option<string> Account = new(
-        $"--{AccountName}"
-    )
+    public static readonly Option<string> Account = new($"--{AccountName}")
     {
         Description = "The name of the App Configuration store (e.g., my-appconfig).",
         Required = true
     };
 
-    public static readonly Option<string> Key = new(
-        $"--{KeyName}"
-    )
+    public static readonly Option<string> Key = new($"--{KeyName}")
     {
         Description = "The name of the key to access within the App Configuration store.",
         Required = true
     };
 
-    public static readonly Option<string> Value = new(
-        $"--{ValueName}"
-    )
+    public static readonly Option<string> Value = new($"--{ValueName}")
     {
         Description = "The value to set for the configuration key.",
         Required = true
     };
 
-    public static readonly Option<string> Label = new(
-        $"--{LabelName}"
-    )
+    public static readonly Option<string> Label = new($"--{LabelName}")
     {
         Description = "The label to apply to the configuration key. Labels are used to group and organize settings.",
         Required = false
     };
 
-    public static readonly Option<string> ContentType = new(
-        $"--{ContentTypeName}"
-    )
+    public static readonly Option<string> ContentType = new($"--{ContentTypeName}")
     {
         Description = "The content type of the configuration value. This is used to indicate how the value should be interpreted or parsed.",
         Required = false
     };
 
-    public static readonly Option<string[]> Tags = new(
-        $"--{TagsName}"
-    )
+    public static readonly Option<string[]> Tags = new($"--{TagsName}")
     {
         Description = "The tags to associate with the configuration key. Tags should be in the format 'key=value'. Multiple tags can be specified.",
         Required = false,
@@ -68,22 +58,15 @@ public static class AppConfigOptionDefinitions
         Required = false
     };
 
-    public static class KeyValueList
+    public static readonly Option<string> KeyFilter = new($"--{KeyFilterName}")
     {
-        public static readonly Option<string> Key = new(
-            $"--{KeyName}"
-        )
-        {
-            Description = "Specifies the key filter, if any, to be used when retrieving key-values. The filter can be an exact match, for example a filter of 'foo' would get all key-values with a key of 'foo', or the filter can include a '*' character at the end of the string for wildcard searches (e.g., 'App*'). If omitted all keys will be retrieved.",
-            Required = false
-        };
+        Description = "Specifies the key filter, if any, to be used when retrieving key-values. The filter can be an exact match, for example a filter of 'foo' would get all key-values with a key of 'foo', or the filter can include a '*' character at the end of the string for wildcard searches (e.g., 'App*'). If omitted all keys will be retrieved.",
+        Required = false
+    };
 
-        public static readonly Option<string> Label = new(
-            $"--{LabelName}"
-        )
-        {
-            Description = "Specifies the label filter, if any, to be used when retrieving key-values. The filter can be an exact match, for example a filter of 'foo' would get all key-values with a label of 'foo', or the filter can include a '*' character at the end of the string for wildcard searches (e.g., 'Prod*'). This filter is case-sensitive. If omitted, all labels will be retrieved.",
-            Required = false
-        };
-    }
+    public static readonly Option<string> LabelFilter = new($"--{LabelFilterName}")
+    {
+        Description = "Specifies the label filter, if any, to be used when retrieving key-values. The filter can be an exact match, for example a filter of 'foo' would get all key-values with a label of 'foo', or the filter can include a '*' character at the end of the string for wildcard searches (e.g., 'Prod*'). This filter is case-sensitive. If omitted, all labels will be retrieved.",
+        Required = false
+    };
 }